+++ Test/UsersAndGroups/Test-ResolveIdentityName.ps1
@@ -17,39 +17,39 @@ function Start-TestFixture
 
 function Test-ShouldResolveBuiltinIdentity
 {
-    $identity = Resolve-IdentityName -Name 'Administrators'
+    $identity = Resolve-CIdentityName -Name 'Administrators'
     Assert-Equal 'BUILTIN\Administrators' $identity
 }
 
 function Test-ShouldResolveNTAuthorityIdentity
 {
-    $identity = Resolve-IdentityName -Name 'NetworkService'
+    $identity = Resolve-CIdentityName -Name 'NetworkService'
     Assert-Equal 'NT AUTHORITY\NETWORK SERVICE' $identity
 }
 
 function Test-ShouldResolveEveryone
 {
-    $identity  = Resolve-IdentityName -Name 'Everyone'
+    $identity  = Resolve-CIdentityName -Name 'Everyone'
     Assert-Equal 'Everyone' $identity
 }
 
 function Test-ShouldNotResolveMadeUpName
 {
-    $fullName = Resolve-IdentityName -Name 'IDONotExist'
+    $fullName = Resolve-CIdentityName -Name 'IDONotExist'
     Assert-NoError
     Assert-Null $fullName
 }
 
 function Test-ShouldResolveLocalSystem
 {
-    Assert-Equal 'NT AUTHORITY\SYSTEM' (Resolve-IdentityName -Name 'localsystem')
+    Assert-Equal 'NT AUTHORITY\SYSTEM' (Resolve-CIdentityName -Name 'localsystem')
 }
 
 function Test-ShouldResolveDotAccounts
 {
     foreach( $user in (Get-User) )
     {
-        $id = Resolve-IdentityName -Name ('.\{0}' -f $user.SamAccountName)
+        $id = Resolve-CIdentityName -Name ('.\{0}' -f $user.SamAccountName)
         Assert-Equal ('{0}\{1}' -f $env:COMPUTERNAME,$user.SamAccountName) $id
     }
 }
@@ -58,13 +58,13 @@ function Test-ShouldResolveBySid
 {
     $id = Resolve-CIdentity -Name 'Administrators' -NoWarn
     Assert-NotNull $id
-    $id = Resolve-IdentityName -Sid $id.Sid.ToString()
+    $id = Resolve-CIdentityName -Sid $id.Sid.ToString()
     Assert-Equal 'BUILTIN\Administrators' $id
 }
 
 function Test-ShouldResolveByUnknownSid
 {
-    $id = Resolve-IdentityName -SID 'S-1-5-21-2678556459-1010642102-471947008-1017'
+    $id = Resolve-CIdentityName -SID 'S-1-5-21-2678556459-1010642102-471947008-1017'
     Assert-Equal $id 'S-1-5-21-2678556459-1010642102-471947008-1017'
     Assert-NoError
 }